pull file path helper into a separate object
since we're going to use it for more than just CLI extraction
